# lite_agent_web/lib/lite_agent_openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "web",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/agent/{id}": {
      "get": {
        "summary": "agent详情",
        "deprecated": false,
        "description": "",
        "tags": [
          "AgentController"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "agent id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataAgentDetailVO",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": {
                        "agent": {
                          "id": "",
                          "userId": "",
                          "workspaceId": "",
                          "name": "",
                          "icon": "",
                          "description": "",
                          "prompt": "",
                          "llmModelId": "",
                          "toolId": [
                            ""
                          ],
                          "status": 0,
                          "shareFlag": false,
                          "createTime": "",
                          "updateTime": ""
                        },
                        "model": {
                          "id": "",
                          "name": "",
                          "baseUrl": "",
                          "apiKey": "",
                          "userId": "",
                          "workspaceId": "",
                          "shareFlag": false,
                          "createTime": "",
                          "updateTime": ""
                        },
                        "toolList": [
                          {
                            "id": "",
                            "userId": "",
                            "workspaceId": "",
                            "name": "",
                            "icon": "",
                            "description": "",
                            "schemaType": 0,
                            "schemaStr": "",
                            "toolStr": "",
                            "apiKey": "",
                            "apiKeyType": "",
                            "shareFlag": false,
                            "createTime": "",
                            "updateTime": ""
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      },
      "delete": {
        "summary": "删除agent",
        "deprecated": false,
        "description": "",
        "tags": [
          "AgentController"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "agent id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Workspace-id",
            "in": "header",
            "description": "工作空间id",
            "required": true,
            "example": "{{workspaceId}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataString",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": ""
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      },
      "put": {
        "summary": "修改agent",
        "deprecated": false,
        "description": "",
        "tags": [
          "AgentController"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "agentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Workspace-id",
            "in": "header",
            "description": "工作空间id",
            "required": true,
            "example": "{{workspaceId}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentUpdateForm",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataAgent",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": {
                        "id": "",
                        "userId": "",
                        "workspaceId": "",
                        "name": "",
                        "icon": "",
                        "description": "",
                        "prompt": "",
                        "llmModelId": "",
                        "toolId": [
                          ""
                        ],
                        "status": 0,
                        "shareFlag": false,
                        "createTime": "",
                        "updateTime": ""
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/v1/agent/add": {
      "post": {
        "summary": "新建agent",
        "deprecated": false,
        "description": "",
        "tags": [
          "AgentController"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Workspace-id",
            "in": "header",
            "description": "工作空间id",
            "required": true,
            "example": "{{workspaceId}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCreateForm",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataAgent",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": {
                        "id": "",
                        "userId": "",
                        "workspaceId": "",
                        "name": "",
                        "icon": "",
                        "description": "",
                        "prompt": "",
                        "llmModelId": "",
                        "toolId": [
                          ""
                        ],
                        "status": 0,
                        "shareFlag": false,
                        "createTime": "",
                        "updateTime": ""
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/v1/agent/list": {
      "get": {
        "summary": "agent列表",
        "deprecated": false,
        "description": "",
        "tags": [
          "AgentController"
        ],
        "parameters": [
          {
            "name": "tab",
            "in": "query",
            "description": "0-全部,1-系统,2-来自分享,3-我的,4-本地agent",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "agent名称",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Workspace-id",
            "in": "header",
            "description": "",
            "required": true,
            "example": "{{workspaceId}}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataListAgentDTO",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": [
                        {
                          "id": "",
                          "userId": "",
                          "workspaceId": "",
                          "name": "",
                          "icon": "",
                          "description": "",
                          "prompt": "",
                          "llmModelId": "",
                          "toolId": [
                            ""
                          ],
                          "status": 0,
                          "shareFlag": false,
                          "createTime": "",
                          "updateTime": ""
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/v1/agent/adminList": {
      "get": {
        "summary": "agent列表(管理)",
        "deprecated": false,
        "description": "",
        "tags": [
          "AgentController"
        ],
        "parameters": [
          {
            "name": "tab",
            "in": "query",
            "description": "0-全部,1-系统,2-来自分享,3-我的",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "agent名称",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Workspace-id",
            "in": "header",
            "description": "",
            "required": true,
            "example": "{{workspaceId}}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataListAgentDTO",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": [
                        {
                          "id": "",
                          "userId": "",
                          "name": "",
                          "icon": "",
                          "description": "",
                          "shareTip": false
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/v1/agent/adminInfo/{id}": {
      "get": {
        "summary": "agent详情(管理)",
        "deprecated": false,
        "description": "",
        "tags": [
          "AgentController"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataAgentDetailVO",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": {
                        "agent": {
                          "id": "",
                          "userId": "",
                          "workspaceId": "",
                          "name": "",
                          "icon": "",
                          "description": "",
                          "prompt": "",
                          "llmModelId": "",
                          "toolIds": [
                            ""
                          ],
                          "status": 0,
                          "shareFlag": false,
                          "temperature": 0,
                          "topP": 0,
                          "maxTokens": 0,
                          "createTime": "",
                          "updateTime": ""
                        },
                        "model": {
                          "id": "",
                          "name": "",
                          "baseUrl": "",
                          "apiKey": "",
                          "userId": "",
                          "workspaceId": "",
                          "shareFlag": false,
                          "createTime": "",
                          "updateTime": ""
                        },
                        "toolList": [
                          {
                            "id": "",
                            "name": "",
                            "icon": "",
                            "description": "",
                            "shareTip": false,
                            "canEdit": false,
                            "canDelete": false,
                            "canRead": false
                          }
                        ],
                        "canEdit": false,
                        "canDelete": false,
                        "canRelease": false
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/v1/agent/release/{id}": {
      "put": {
        "summary": "发布",
        "deprecated": false,
        "description": "",
        "tags": [
          "AgentController"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "agent id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Workspace-id",
            "in": "header",
            "description": "工作空间id",
            "required": true,
            "example": "{{workspaceId}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataString",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": ""
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/v1/agent/generateApiKey/{id}": {
      "post": {
        "summary": "生成agent ApiKey",
        "deprecated": false,
        "description": "",
        "tags": [
          "AgentController"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "agent id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Workspace-id",
            "in": "header",
            "description": "工作空间id",
            "required": true,
            "example": "{{workspaceId}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataAgentApiKey",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": {
                        "apiKey": "",
                        "apiUrl": ""
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/v1/agent/resetSequence/{agentId}": {
      "post": {
        "summary": "重置方法序列",
        "deprecated": false,
        "description": "",
        "tags": [
          "AgentController"
        ],
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "认证 Token",
            "required": true,
            "example": "Bearer {{token}}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Workspace-id",
            "in": "header",
            "description": "工作空间id",
            "required": true,
            "example": "{{workspaceId}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDataString",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "message": "",
                      "data": ""
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResponseDataAgentDetailVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "响应信息"
          },
          "data": {
            "$ref": "#/components/schemas/AgentDetailVO",
            "description": "响应数据对象"
          }
        }
      },
...
      "AgentUpdateForm": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "agent名称"
          },
          "icon": {
            "type": "string",
            "description": "图标url"
          },
          "description": {
            "type": "string",
            "description": "描述,可空"
          },
          "prompt": {
            "type": "string",
            "description": "提示词"
          },
          "llmModelId": {
            "type": "string",
            "description": "关联的大模型id"
          },
          "toolIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "工具"
          },
          "temperature": {
            "type": "number",
            "description": "温度值",
            "minimum": 0,
            "maximum": 1,
            "x-apifox-mock": "@float(0,1,2)"
          },
          "topP": {
            "type": "number",
            "description": "概率抽样的 p 值",
            "minimum": 0,
            "maximum": 1,
            "x-apifox-mock": "@float(0,1,2)"
          },
          "maxTokens": {
            "type": "integer",
            "description": "最大 token 数",
            "minimum": 1,
            "x-apifox-mock": "@integer(1)"
          },
          "type": {
            "type": "integer",
            "description": "agent类型",
            "enum": [
              0,
              1,
              2
            ],
            "x-apifox-enum": [
              {
                "value": 0,
                "name": "GENERAL",
                "description": "普通"
              },
              {
                "value": 1,
                "name": "DISTRIBUTE",
                "description": "分发"
              },
              {
                "value": 2,
                "name": "REFLECTION",
                "description": "反思"
              }
            ],
            "default": 0
          },
          "mode": {
            "type": "integer",
            "description": "执行模式",
            "enum": [
              0,
              1,
              2
            ],
            "x-apifox-enum": [
              {
                "value": 0,
                "name": "PARALLEL",
                "description": "并行"
              },
              {
                "value": 1,
                "name": "SERIAL",
                "description": "串行"
              },
              {
                "value": 2,
                "name": "REJECT",
                "description": "拒绝"
              }
            ],
            "default": 0
          },
          "subAgentIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "子agent(普通,分发,分思) id"
          },
          "functionList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentFunction",
              "description": "com.litevar.agent.base.entity.Agent.AgentFunction"
            },
            "description": "方法列表"
          },
          "sequence": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "方法执行顺序(functionId)"
          },
          "datasetIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "数据集id"
          }
        },
        "required": [
          "name"
        ]
      },
      "AgentCreateForm": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "agent名称"
          },
          "icon": {
            "type": "string",
            "description": "图标url"
          },
          "description": {
            "type": "string",
            "description": "描述,可空"
          }
        },
        "required": [
          "name"
        ]
      },
      "ResponseDataAgent": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "响应信息"
          },
          "data": {
            "$ref": "#/components/schemas/Agent",
            "description": ""
          }
        }
      },
...
      "ResponseDataString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "响应信息"
          },
          "data": {
            "type": "string",
            "description": "响应数据对象"
          }
        }
      }
    },
    "securitySchemes": {}
  },
  "servers": []
}